Data.Semiring.Class

Description

Semirings, with and without multiplicative identities.

Semirings

class Semigroup r => Semiring r where

A Semiring r is an abstract algebraic structure consisting of a commutative Semigroup and an associative operator ><, with the additional constraints that >< distributes over <>. Additionally, if s is a Monoid, zero is the absorping element of ><.

Laws:

Associativity of <> (the Semigroup law):

  a <> (b <> c) = (a <> b) <> c

Identity of <> (the Monoid law, if r is a Monoid):

  zero <> a    = a
  a    <> zero = a

Commutativity of <>:

  a <> b = b <> a

Associativity of ><:

  a >< (b >< c) = (a >< b) >< c

Distributivity of >< over <>:

  a >< (b <> c) = (a >< b) <> (a >< c)
  (a <> b) >< c = (a >< c) <> (b >< c)

Absorption of >< by zero (if r is a Monoid):

  zero >< a    = zero
  a    >< zero = zero

Minimal complete definition

(><)

Methods

(><) :: r -> r -> r infixr 7

zero :: Monoid m => m

The identity element of some Monoid m.

If m is additionally a Semiring, this is its multiplicative absorping element.

Unital semirings

class Semiring r => Unital r where

Unital Semirings additionally have a multiplicative identity, one.

Laws:

Identity of ><:

  one >< a   = a
  a   >< one = a

Minimal complete definition

one

Methods

one :: r